feature(UserFeedback): files organization
