+++ DidIFloss/Service/UserFeedback/UserFeedbackManagerProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  UserFeedbackManagerProtocol.swift
+//  DidIFloss
+//
+//  Created by Lucas Migge on 22/02/24.
+//
+
+import Foundation
+
+protocol UserFeedbackManagerProtocol {
+    func vibrateCelebration()
+    func vibrateDeletion()
+    
+}
